X11 sucks
Apparently, no matter how hard you set the clipboard, the clipboard clears
itself as soon as your program exits. This is because the guy who wrote X11
never knew the love of a woman and chose to seize the Rheingold even though
everyone told him not to, and now we all have to suffer for it.